requirements.txt
python-telegram-bot==20.3
nudenet==3.4.2
tensorflow==2.10.0
opencv-python-headless==4.7.0.72
pillow==9.5.0
python-dotenv==1.0.0
numpy==1.23.5